+++ src/screens/account_details/components/nonce_account/styles.ts
@@ -6,17 +6,12 @@ export const useStyles = () => {
       return ({
         overview: {
           [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
+            gridColumn: '1 / 2',
           },
         },
         tokens: {
           [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
-          },
-        },
-        transactions: {
-          [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
+            gridColumn: '1 / 2',
           },
         },
       });