+++ src/screens/account_details/components/token_details_account/index.tsx
@@ -2,9 +2,6 @@ import React from 'react';
 import {
   Overview,
 } from './components';
-import {
-  Transactions,
-} from '..';
 import { useStyles } from './styles';
 import { useTokenDetailAccount } from './hooks';
 
@@ -13,10 +10,7 @@ const TokenAccount = () => {
   const { state } = useTokenDetailAccount();
 
   return (
-    <>
-      <Overview className={classes.overview} overview={state.overview} />
-      <Transactions className={classes.transactions} />
-    </>
+    <Overview className={classes.overview} overview={state.overview} />
   );
 };