+++ src/screens/account_details/components/token_details_account/styles.ts
@@ -6,22 +6,7 @@ export const useStyles = () => {
       return ({
         overview: {
           [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
-          },
-        },
-        accounts: {
-          [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
-          },
-        },
-        tokens: {
-          [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
-          },
-        },
-        transactions: {
-          [theme.breakpoints.up('lg')]: {
-            gridColumn: '1 / 3',
+            gridColumn: '1 / 2',
           },
         },
       });