+++ src/screens/account_details/components/vote_account/styles.ts
@@ -4,12 +4,17 @@ export const useStyles = () => {
   const styles = makeStyles(
     (theme) => {
       return ({
+        validator: {
+          [theme.breakpoints.up('lg')]: {
+            gridColumn: '1 / 3',
+          },
+        },
         overview: {
           [theme.breakpoints.up('lg')]: {
             gridColumn: '1 / 3',
           },
         },
-        accounts: {
+        skipRate: {
           [theme.breakpoints.up('lg')]: {
             gridColumn: '1 / 3',
           },